requirements.txt
# Core ML and Data Processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter and Notebooks
ipykernel>=6.0.0
jupyter>=1.0.0
notebook>=6.4.0

# Data Storage
pyarrow>=10.0.0
fastparquet>=0.8.0

# MLflow and Experiment Tracking
mlflow>=2.0.0
xgboost>=1.6.0

# API Framework
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0

# Configuration
pyyaml>=6.0

# AI/LLM Integration
google-generativeai>=0.3.0

# State Management
redis>=4.5.0

# Utilities
joblib>=1.2.0
python-multipart>=0.0.5
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
flake8>=5.0.0
mypy>=1.0.0
black>=22.0.0
isort>=5.10.0

# Security
bandit>=1.7.0
safety>=2.0.0

# Production
gunicorn>=20.1.0